Shuffling from atomic positions works now
