+++ mala/datahandling/data_handler.py
@@ -1,7 +1,6 @@
 """DataHandler class that loads and scales data."""
 
 import os
-import tempfile
 
 import numpy as np
 import torch
@@ -170,21 +169,6 @@ class DataHandler(DataHandlerBase):
         self.output_data_scaler.reset()
         super(DataHandler, self).clear_data()
 
-    def delete_temporary_inputs(self):
-        """
-        Delete temporary data files.
-
-        These may have been created during a training or testing process
-        when using atomic positions for on-the-fly calculation of descriptors
-        rather than precomputed data files.
-        """
-        if get_rank() == 0:
-            for snapshot in self.parameters.snapshot_directories_list:
-                if snapshot.temporary_input_file is not None:
-                    if os.path.isfile(snapshot.temporary_input_file):
-                        os.remove(snapshot.temporary_input_file)
-        barrier()
-
     # Preparing data
     ######################
 
@@ -348,31 +332,6 @@ class DataHandler(DataHandlerBase):
         return self.parameters.snapshot_directories_list[
             snapshot_number
         ].calculation_output
-
-    def __calculate_temporary_inputs(self, snapshot: Snapshot):
-        if snapshot.temporary_input_file is not None:
-            if not os.path.isfile(snapshot.temporary_input_file):
-                snapshot.temporary_input_file = None
-
-        if snapshot.temporary_input_file is None:
-            snapshot.temporary_input_file = tempfile.NamedTemporaryFile(
-                delete=False,
-                prefix=snapshot.input_npy_file.split(".")[0],
-                suffix=".in.npy",
-                dir=snapshot.input_npy_directory,
-            ).name
-            tmp, grid = self.descriptor_calculator.calculate_from_json(
-                os.path.join(
-                    snapshot.input_npy_directory,
-                    snapshot.input_npy_file,
-                )
-            )
-            if self.parameters._configuration["mpi"]:
-                tmp = self.descriptor_calculator.gather_descriptors(tmp)
-
-            if get_rank() == 0:
-                np.save(snapshot.temporary_input_file, tmp)
-            barrier()
 
     # Debugging
     ######################
@@ -637,7 +596,7 @@ class DataHandler(DataHandlerBase):
                     # If the input for the descriptors is actually a JSON
                     # file then we need to calculate the descriptors.
                     if snapshot.snapshot_type == "json+numpy":
-                        self.__calculate_temporary_inputs(snapshot)
+                        self._calculate_temporary_inputs(snapshot)
                         file = snapshot.temporary_input_file
                 else:
                     file = os.path.join(
@@ -768,14 +727,14 @@ class DataHandler(DataHandlerBase):
                 # snapshot types, this has to be done here.
                 if snapshot.snapshot_function == "va":
                     if snapshot.snapshot_type == "json+numpy":
-                        self.__calculate_temporary_inputs(snapshot)
+                        self._calculate_temporary_inputs(snapshot)
 
                     self.validation_data_sets[0].add_snapshot_to_dataset(
                         snapshot
                     )
                 if snapshot.snapshot_function == "te":
                     if snapshot.snapshot_type == "json+numpy":
-                        self.__calculate_temporary_inputs(snapshot)
+                        self._calculate_temporary_inputs(snapshot)
 
                     self.test_data_sets[0].add_snapshot_to_dataset(snapshot)
 
@@ -822,7 +781,7 @@ class DataHandler(DataHandlerBase):
                         )
                     )
                     if snapshot.snapshot_type == "json+numpy":
-                        self.__calculate_temporary_inputs(snapshot)
+                        self._calculate_temporary_inputs(snapshot)
 
                 if snapshot.snapshot_function == "te":
                     self.test_data_sets.append(
@@ -840,7 +799,7 @@ class DataHandler(DataHandlerBase):
                         )
                     )
                     if snapshot.snapshot_type == "json+numpy":
-                        self.__calculate_temporary_inputs(snapshot)
+                        self._calculate_temporary_inputs(snapshot)
 
         else:
             if self.nr_training_data != 0:
@@ -940,7 +899,7 @@ class DataHandler(DataHandlerBase):
                             )
                         )
                     elif snapshot.snapshot_type == "json+numpy":
-                        self.__calculate_temporary_inputs(snapshot)
+                        self._calculate_temporary_inputs(snapshot)
                         tmp = self.descriptor_calculator.read_from_numpy_file(
                             snapshot.temporary_input_file,
                             units=snapshot.input_units,
@@ -1040,7 +999,7 @@ class DataHandler(DataHandlerBase):
                     snapshot.snapshot_function == "tr"
                     and snapshot.snapshot_type == "json+numpy"
                 ):
-                    self.__calculate_temporary_inputs(snapshot)
+                    self._calculate_temporary_inputs(snapshot)
         else:
             printout(
                 "Data scalers already initilized, loading data to RAM.",
